requirement.txt
appdirs==1.4.4
astroid==2.5.1
buildozer==1.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Cython==0.29.19
distlib==0.3.1
docutils==0.16
ffpyplayer==4.3.2
filelock==3.0.12
gTTS==2.2.2
gTTS-token==1.1.4
idna==2.10
importlib-metadata==3.7.3
install==1.3.4
isort==5.8.0
Jinja2==2.11.3
Kivy==2.0.0
Kivy-Garden==0.1.4
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pep517==0.6.0
pexpect==4.8.0
Pillow==8.1.2
ptyprocess==0.7.0
Pygments==2.8.1
pylint==2.7.2
PyPDF2==1.26.0
python-for-android==2020.6.2
pytoml==0.1.21
requests==2.25.1
sh==1.14.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
wrapt==1.12.1
zipp==3.4.1
